Add level2(map2)
Level 2 was added.
The bitmap variable is a public variable in map.h, so it is accessed by map.bitmap.
And I changed the reset function so that it can be initialized to level according to the level2 and level3.
Also, I changed the gameover function so that when we ate all the dots in level2, we could go to level3.
Also, I changed the part of the display that draws the right screen so that the coordinate axes are scaled according to the level. As a result, the level 2 is 13by13, and the level 3 is the existing 15by15.